requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
tqdm>=4.65.0
reportlab>=4.0.0
streamlit>=1.28.0
